修复了级联删除的bug(*^__^*)

+++ src/main/java/com/zxd/bbs/dao/ReplyMapper.java
@@ -13,6 +13,6 @@ public interface ReplyMapper {
 	
 	int insert(Reply reply);
 	
-	int deleteById(long id);
+	int deleteById(Integer id);
 	
 }
